Rearange special forms and pseudo keywords
Now that "special forms" are a defined concept in the language
documentation some of the symbol names in the special form and pseudo
keywords are in the wrong respective category.
This commit fixes that and sorts the symbols in alphabetical order.

I also took the decision that the builtin macros are special forms,
since they share the defining property that you can't stop them from
evaluating their arguments (if they do that).